requirements/runtime.txt
matplotlib
mmcv>=0.6.2
numpy
# need older pillow until torchvision is fixed
Pillow<=6.2.2
six
terminaltables
torch>=1.3
torchvision
scikit-image
tqdm
einops
